There are a lots
The list works by basically shifting a pointer and then making the new
node that it points to take over the job of the now shifted one. This
adds an element. Slight variations happen for the front, and all of
them add one to the size of the list, which is why they are more
dynamic than arrays.